Explicit cast warning-error

+++ test/util/util.c
@@ -699,7 +699,8 @@ _state_need_kms (_state_machine_t *state_machine, bson_error_t *error)
          }
 
          mongocrypt_binary_destroy (http_reply);
-         http_reply = mongocrypt_binary_new_from_data (buf, read_ret);
+         http_reply =
+            mongocrypt_binary_new_from_data (buf, (uint32_t) read_ret);
          if (!mongocrypt_kms_ctx_feed (kms_ctx, http_reply)) {
             _kms_ctx_check_error (kms_ctx, error, true);
             goto fail;
